Update for new huggingface_hub version and simplify implementation
